Extract rich markdown review annotation ranges
